refactor: extract Unroll functionality into a dedicated module
